backend/requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
fonttools==4.39.3
grad-cam==1.4.6
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sympy==1.11.1
threadpoolctl==3.1.0
timm==0.6.13
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
ttach==0.0.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
